gFeatures 2.5.5 - The largest Player API Patch to date.
Highly recommended to update to this patch!

+++ src/net/genesishub/gFeatures/Feature/gWarsSuite/Multiplayer/Damage.java
@@ -13,41 +13,35 @@ public class Damage {
 	Source s = new Source();
 	Statistics stats = new Statistics();
 	public void onEntityDamage(EntityDamageByEntityEvent event){
-		Bukkit.getLogger().info("Regular Kill.");
 		if((OrangeTeam.hasPlayer((Player) event.getDamager()) &&  OrangeTeam.hasPlayer((Player) event.getEntity())) || (BlueTeam.hasPlayer((Player) event.getDamager()) &&  BlueTeam.hasPlayer((Player) event.getEntity()))){
 			event.setCancelled(true);
-			Bukkit.getLogger().info("Damage averted.");
 		}
 		else{
 			Player death = (Player) event.getEntity();
 			Damageable d = (Damageable) death;
 			int health = (int) d.getHealth();
 			double damage = event.getDamage();
-			//if(health - damage <= 0){
+			if(health - damage <= 0){
 				stats.addDeaths((Player)event.getEntity(), 1);
 				stats.addKill((Player)event.getDamager(), 1);
 				s.flushAll();
-				Bukkit.getLogger().info("Done.");
-			//}
+			}
 		}
 	}
 	public void onWeaponDamage(WeaponDamageEntityEvent event){
-		Bukkit.getLogger().info("Weapon Kill");
 		if((OrangeTeam.hasPlayer((Player) event.getPlayer()) &&  OrangeTeam.hasPlayer((Player) event.getVictim())) || (BlueTeam.hasPlayer((Player) event.getPlayer()) &&  BlueTeam.hasPlayer((Player) event.getVictim()))){
 			event.setCancelled(true);
-			Bukkit.getLogger().info("Damage averted.");
 		}
 		else{
 			Player death = (Player) event.getVictim();
 			Damageable d = (Damageable) death;
 			int health = (int) d.getHealth();
 			double damage = event.getDamage();
-			//if(health - damage <= 0){
+			if(health - damage <= 0){
 				stats.addDeaths((Player)event.getVictim(), 1);
 				stats.addKill((Player)event.getPlayer(), 1);
 				s.flushAll();
-				Bukkit.getLogger().info("Done.");
-			//}
+			}
 		}
 	}
 }
